[RHODA][UI]Adding test for adding duplicate provider account name
This PR adds a new test suite which will be used for test cases with generic use.
The test suite contains a test case where a user tries to import a new provider account with a provider name that already exists.
The PR includes the following:
1. A new test suite file: generic_tests.robot
2. An update to the file: resources/keywords/import_provider_account.resource for adding necesary keywords for the test case
3. An update to the file: resources/object_repo/import_dbaasinventory_obj.resource for adding the relevant web element

This PR resolves Jira ticket DBAAS-693